Fix ack_len
Fixes some routers not being able to connect. ack_len tended to be set to 0xFF on broadcast packets, despite not having any ack data, which would result in a negative packet size if the packet was shorter than 0xFF. This would cause it to not process the packet correctly; the fact that larger packets processed correctly is somewhat of a mystery to me as they theoretically should've bailed early as well, but it's likely it's just a certain threshold before it stops working.